internal: Update flush counter to work around bug
There is an off by one bug in the flush log delay that is causing context to be lost under specific conditions. This change works around that bug by adding an additional flush count when a delayed flush is required.

Future changes will move to using the newer local log file contract instead.